Changes in lab view, added counties table and controller

+++ app/Http/Controllers/LabController.php
@@ -12,6 +12,8 @@ use App\Models\LabAffiliation;
 use App\Models\LabType;
 use App\Models\AuditType;
 use App\Models\Country;
+use App\Models\County;
+use App\Models\Review;
 use Response;
 use Auth;
 use Session;
@@ -27,6 +29,7 @@ class LabController extends Controller {
 	{
 		//	Get all labs
 		$labs = Lab::all();
+		// dd($labs);
 		//	Get audit types
 		$auditTypes = AuditType::all();
 		/*Default audit type*/
@@ -41,15 +44,15 @@ class LabController extends Controller {
 	 */
 	public function create()
 	{
-		//	Get all countries
-		$countries = Country::lists('name', 'id');
+		//	Get all counties
+		$counties = County::lists('name', 'id');
 		//	Get all lab levels
 		$labLevels = LabLevel::lists('name', 'id');
 		//	Get all lab affiliations
 		$labAffiliations = LabAffiliation::lists('name', 'id');
 		//	Get all lab types
 		$labTypes = LabType::lists('name', 'id');
-		return view('lab.lab.create', compact('countries', 'labLevels', 'labAffiliations', 'labTypes'));
+		return view('lab.lab.create', compact('counties', 'labLevels', 'labAffiliations', 'labTypes'));
 	}
 
 	/**
@@ -60,20 +63,18 @@ class LabController extends Controller {
 	public function store(LabRequest $request)
 	{
 		$lab = new Lab;
-		$lab->lab_type_id = $request->lab_type;
-        $lab->name = $request->name;
+		$lab->name = $request->name;
 		$lab->lab_number = $request->number;
-		$lab->address = $request->address;
-		$lab->postal_code = $request->postal_code;
-		$lab->city = $request->city;
-		$lab->state = $request->state;
-        $lab->country_id = $request->country;
-		$lab->fax = $request->fax;
+		$lab->lab_type_id = $request->lab_type;
+       	$lab->lab_level_id = $request->lab_level;
+        $lab->lab_affiliation_id = $request->lab_affiliation;
+        $lab->address = $request->address;
+		$lab->postal_address = $request->postal_address;
+		$lab->county_id = $request->county_id;
+		$lab->subcounty = $request->state;
 		$lab->telephone = $request->telephone;
 		$lab->email = $request->email;
-		$lab->lab_level_id = $request->lab_level;
-        $lab->lab_affiliation_id = $request->lab_affiliation;
-        $lab->user_id = Auth::user()->id;
+		 $lab->user_id = Auth::user()->id;
         $lab->save();
         $url = session('SOURCE_URL');
 
@@ -105,9 +106,9 @@ class LabController extends Controller {
 		//	Get lab
 		$lab = Lab::find($id);
 		//	Get all countries
-		$countries = Country::lists('name', 'id');
+		$counties = County::lists('name', 'id');
 		//	Get initially selected country
-		$country = $lab->country_id;
+		$county = $lab->county_id;
 		//	Get all lab levels
 		$labLevels = LabLevel::lists('name', 'id');
 		//	Get initially selected lab level
@@ -121,7 +122,7 @@ class LabController extends Controller {
 		//	Get initially selected lab type
 		$labType = $lab->lab_type_id;
 
-        return view('lab.lab.edit', compact('lab', 'countries', 'labLevels', 'labAffiliations', 'labTypes', 'country', 'labLevel', 'labAffiliation', 'labType'));
+        return view('lab.lab.edit', compact('lab', 'counties', 'labLevels', 'labAffiliations', 'labTypes', 'county', 'labLevel', 'labAffiliation', 'labType'));
 	}
 
 	/**
@@ -134,18 +135,15 @@ class LabController extends Controller {
 	{
 		$lab = Lab::findOrFail($id);
         $lab->lab_type_id = $request->lab_type;
+		$lab->lab_level_id = $request->lab_level;
+        $lab->lab_affiliation_id = $request->lab_affiliation;
         $lab->name = $request->name;
 		$lab->lab_number = $request->number;
 		$lab->address = $request->address;
-		$lab->postal_code = $request->postal_code;
-		$lab->city = $request->city;
-		$lab->state = $request->state;
-        $lab->country_id = $request->country;
-		$lab->fax = $request->fax;
+		$lab->postal_address = $request->postal_address;
+		$lab->county_id = $request->county_id;
 		$lab->telephone = $request->telephone;
 		$lab->email = $request->email;
-		$lab->lab_level_id = $request->lab_level;
-        $lab->lab_affiliation_id = $request->lab_affiliation;
         $lab->user_id = Auth::user()->id;;
         $lab->save();
         $url = session('SOURCE_URL');
@@ -160,10 +158,22 @@ class LabController extends Controller {
 	 * @return Response
 	 */
 	public function delete($id)
-	{
+	{  
 		$lab= Lab::find($id);
-		$lab->delete();
-		return redirect('lab')->with('message', 'Lab deleted successfully.');
+		$labReviewInUse = Review::where('lab_id', '=', $id)->first();
+
+		if (empty($labReviewInUse)) 
+		{
+			// no review created for the lab
+			$lab->delete();
+			return redirect('lab')->with('message', 'Lab deleted successfully.');
+
+		} else{
+
+			return redirect('lab')->with('message', 'Lab is in use.');
+
+		}
+
 	}
 	public function destroy($id)
 	{
